Kernel32: Enable beeping.
Previously, kernel32-sys was unsuitable for use in cars. Also unsuitable for prank programs that require a beep that bypasses a system mute.

I take no responsibility for any future use that causes data centre owners to require earmuffs or deafness as a job requirement.

I wonder what that would be like, a room full of constantly beeping servers...