Reverse order of overloaded COM methods in vtable (#524)
Puts overloaded COM methods in reverse order to follow MSVC ABI
convention. Also changes the naming convention to use an underscore
and a 1-based number for all overrides but the first, from the
current convention which is not consistent.

Tested with animations using the DirectComposition interfaces, which
are most of the changes. The main other cluster is font management
in DirectWrite.