add support for aarch64-pc-windows-msvc (#677)
* add aarch64 conditionals and types

Most of the additions are where `_WIN64` was being checked, in which
case we add a case next to the `target_arch = "x86_64"` checks.  Other
headers, particularly winnt.h, require custom AArch64-specific types.

* use target_pointer_width rather than explicit target_arch selections

This style is shorter, and corresponds more closer to the headers
depending on _WIN64 and the like.